Use root of --url as <base>.  Load reliably.
Also, disable the cache by using "incognito" mode so that we always
load the latest resources.

The new approach to loading ensures that DREI resources are loaded
from Tauri, but that other resources at relative links are loaded
relative to the root of --url, if it is supplied.  The rewriting
approach we used before was fragile.